Reorganize ne_NP person test
